refactor: make `semantic` package public
