+++ pkg/semantic/compare.go
