+++ pkg/semantic/fixtures/all-versions.txt
